refactor(desktop): organize terminal hooks
